add: mymapを削除して、新たにrecordのフォルダを作成し、そこに移動しました
